forces/pyproject.toml
[tool.poetry]
authors = ["Product Team <development@fluidattacks.com>"]
description = "Forces is simple and powerful, it allows validating the status of vulnerabilities within integrates."
license = "MPL-2.0"
name = "forces"
packages = [
  { include = "forces", from = "." }
]
readme = "./README.md"
repository = "https://gitlab.com/fluidattacks/integrates/-/tree/master/forces"
version = "1.0.0"

[tool.poetry.dependencies]
aiohttp = "3.6.2"
click = "7.1.2"
gql = "3.0.0a1"
oyaml = "0.9"
python-jose = "3.1.0"
python = "^3.8"

[tool.poetry.dev-dependencies]
prospector = "1.2.0"
mypy = "0.770"
pytest = "5.4.3"
pytest-asyncio = "0.14.0"
pytest-cov = "2.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
forces = 'forces.cli:main'
